change file name and include directory structure
