+++ planning/autoware_trajectory_ranker/include/autoware/trajectory_ranker/metrics/time_to_collision_metric.hpp
@@ -12,28 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE_TRAJECTORY_RANKER__METRICS__LATERAL_DEVIATION_METRIC_HPP_
-#define AUTOWARE_TRAJECTORY_RANKER__METRICS__LATERAL_DEVIATION_METRIC_HPP_
+#ifndef AUTOWARE__TRAJECTORY_RANKER__METRICS__TIME_TO_COLLISION_METRIC_HPP_
+#define AUTOWARE__TRAJECTORY_RANKER__METRICS__TIME_TO_COLLISION_METRIC_HPP_
 
-#include "autoware_trajectory_ranker/interface/metrics_interface.hpp"
+#include "autoware/trajectory_ranker/interface/metrics_interface.hpp"
 
 #include <memory>
 
 namespace autoware::trajectory_ranker::metrics
 {
 
-class LateralDeviation : public MetricInterface
+class TimeToCollision : public MetricInterface
 {
 public:
-  LateralDeviation() : MetricInterface("LateralDeviation") {}
+  TimeToCollision() : MetricInterface("TimeToCollision") {}
 
   void evaluate(
     const std::shared_ptr<autoware::trajectory_ranker::DataInterface> & result,
     const double max_value) const override;
 
-  bool is_deviation() const override { return true; }
+  bool is_deviation() const override { return false; }
 };
 
 }  // namespace autoware::trajectory_ranker::metrics
 
-#endif  // AUTOWARE_TRAJECTORY_RANKER__METRICS__LATERAL_DEVIATION_METRIC_HPP_
+#endif  // AUTOWARE__TRAJECTORY_RANKER__METRICS__TIME_TO_COLLISION_METRIC_HPP_
