+++ planning/autoware_trajectory_ranker/include/autoware/trajectory_ranker/utils.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE_TRAJECTORY_RANKER__UTILS_HPP_
-#define AUTOWARE_TRAJECTORY_RANKER__UTILS_HPP_
+#ifndef AUTOWARE__TRAJECTORY_RANKER__UTILS_HPP_
+#define AUTOWARE__TRAJECTORY_RANKER__UTILS_HPP_
 
 #include <rclcpp/duration.hpp>
 
@@ -56,4 +56,4 @@ TrajectoryPoint calc_interpolated_point(
 
 }  // namespace autoware::trajectory_ranker::utils
 
-#endif  // AUTOWARE_TRAJECTORY_RANKER__UTILS_HPP_
+#endif  // AUTOWARE__TRAJECTORY_RANKER__UTILS_HPP_
