+++ planning/autoware_trajectory_ranker/src/evaluation.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware_trajectory_ranker/evaluation.hpp"
+#include "autoware/trajectory_ranker/evaluation.hpp"
 
 #include <rclcpp/logging.hpp>
