+++ planning/autoware_trajectory_ranker/src/metrics/distance_metric.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware_trajectory_ranker/metrics/distance_metric.hpp"
+#include "autoware/trajectory_ranker/metrics/distance_metric.hpp"
 
 #include <autoware/motion_utils/trajectory/trajectory.hpp>
