+++ planning/autoware_trajectory_ranker/src/metrics/lateral_acceleration_metric.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware_trajectory_ranker/metrics/lateral_acceleration_metric.hpp"
+#include "autoware/trajectory_ranker/metrics/lateral_acceleration_metric.hpp"
 
 #include <tf2/utils.h>
