+++ planning/autoware_trajectory_ranker/src/metrics/longitudinal_jerk_metric.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware_trajectory_ranker/metrics/longitudinal_jerk_metric.hpp"
+#include "autoware/trajectory_ranker/metrics/longitudinal_jerk_metric.hpp"
 
 #include <algorithm>
 #include <cmath>
