+++ planning/autoware_trajectory_ranker/src/metrics/steering_consistency_metric.cpp
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware_trajectory_ranker/metrics/steering_consistency_metric.hpp"
+#include "autoware/trajectory_ranker/metrics/steering_consistency_metric.hpp"
 
-#include "autoware_trajectory_ranker/metrics/metrics_utils.hpp"
+#include "autoware/trajectory_ranker/metrics/metrics_utils.hpp"
 
 #include <tf2/utils.h>
